requirements.txt
numpy
pathlib
torch
pandas
torchvision
matplotlib
pillow 
wandb
